requirements.txt
falcon==3.1.3
gitdb==4.0.11
GitPython==3.1.40
gunicorn==21.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
smmap==5.0.1
